feat(plasma-new-hope): resolve circular dependency in engine dir
