+++ packages/plasma-new-hope/src/engines/common.tsx
@@ -3,52 +3,12 @@ import React from 'react';
 import { _component } from './linaria';
 import type {
     ComponentConfig,
-    HTMLAnyAttributes,
     HTMLTagList,
     PropsType,
     Variants,
     HTMLAttributesWithoutOnChange,
     HTMLAttributesWithoutOnChangeAndDefaultValue,
 } from './types';
-
-//
-// Тип HTMLAttributesOmitOnChange требуется чтобы использовать компонент с кастомным пропсом onChange
-//
-
-export const getStaticVariants = (config: ComponentConfig) => {
-    if (!config.variations) {
-        return [];
-    }
-    const res = [];
-    const { variations } = config;
-
-    // eslint-disable-next-line guard-for-in
-    for (const key in variations) {
-        const { css } = variations[key];
-        css && res.push(css);
-    }
-
-    return res;
-};
-
-export const getDynamicVariants = (config: ComponentConfig) => {
-    return (props: HTMLAnyAttributes) => {
-        const res = [];
-
-        // eslint-disable-next-line guard-for-in
-        for (const key in config.variations) {
-            if (key in props) {
-                const variant = config.variations[key];
-
-                const css = variant[props[key]];
-                // no css for { modifier: true }
-                css && Array.isArray(css) ? res.push(...css) : res.push(css);
-            }
-        }
-
-        return res;
-    };
-};
 
 export const mergeConfig = <
     Tag extends HTMLTagList,
